JUnit Basic Navigation Test
Using JUnit I navigate to a web page,then navigate to a new page and set the current URL then print the string that is was saved to then do basic back and forward navigation as well as a page refresh and a refresh based on the last visited URL. I also print the page source, print validation at each step. Includes a setup and tear down.